submit posts test changed in order to account for the card order
