+++ spec/features/user_can_submit_posts_spec.rb
@@ -19,12 +19,12 @@ RSpec.feature "Timeline", type: :feature do
   scenario "User email displayed with a post" do
     user_signs_up
     user_makes_a_post
-    expect(page).to have_content("jordan@matt.com: Hello, world!")
+    expect(page).to have_content("Hello, world! jordan@matt.com")
   end
 
-  scenario "Time is displayed with post" do 
+  scenario "Time is displayed with post" do
     user_signs_up
     user_makes_a_post
-    expect(page).to have_content("jordan@matt.com: Hello, world! less than a minute ago")
-  end  
+    expect(page).to have_content("Hello, world! jordan@matt.com less than a minute ago")
+  end
 end
